Trait astar_primitives::dapp_staking::SmartContractHandle

pub trait SmartContractHandle<AccountId> {
    // Required methods
    fn evm(address: H160) -> Self;
    fn wasm(address: AccountId) -> Self;
}

Trait defining the interface for dApp staking smart contract types handler.

It can be used to create a representation of the specified smart contract instance type.

Required Methods

fn evm(address: H160) -> Self

Create a new smart contract representation for the specified EVM address.

fn wasm(address: AccountId) -> Self

Create a new smart contract representation for the specified Wasm address.

Object Safety

This trait is not object safe.