astar_primitives

Module evm

EVM primitives.

Structs

EVMFungibleAdapterWrapper

Wrapper around the EvmFungibleAdapter from the pallet-evm.

EvmRevertCodeHandler

Handler for automatic revert code registration.

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

HashedDefaultMappings

Mappings derieved from hashing the original address

U256

Little-endian large integer type 256-bit unsigned integer.

Enums

UnifiedAddress

Constants

EVM_REVERT_CODE

Revert opt code. It’s inserted at the precompile addresses, to make them functional in EVM.

Traits

Hasher

Trait describing an object that can hash a slice of bytes. Used to abstract other types over the hashing algorithm. Defines a single hash method and an Out associated type with the necessary bounds.

UnifiedAddressMapper

Mapping between Native and EVM Addresses

Type Aliases

EvmAddress