Type Alias astar_primitives::AccountId

pub type AccountId = <<Signature as Verify>::Signer as IdentifyAccount>::AccountId;

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.