Crate astar_primitives

Core Astar types.

These core Astar types are used by the Shiden, Shibuya, Astar and Local runtime.

Modules

dapp_staking

dApp staking & inflation primitives.

ethereum_checked

Checked Ethereum transaction primitives.

evm

EVM primitives.

oracle

Oracle & price primitives.

precompiles

Precompiles

testing

Useful primitives for testing.

xcm

XCM Primitives

xvm

XVM primitives.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AssetId

Id used for identifying assets.

Balance

Balance of an account.

Block

Block type.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Header

Block header type as expected by this runtime.

Nonce

Index of a transaction in the chain.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.