Module astar_primitives::xcm

XCM Primitives

Overview

Collection of common XCM primitives used by runtimes.

AssetLocationIdConverter - conversion between local asset Id and cross-chain asset multilocation

FixedRateOfForeignAsset - weight trader for execution payment in foreign asset

ReserveAssetFilter - used to check whether asset/origin are a valid reserve location

XcmFungibleFeeHandler - used to handle XCM fee execution fees

Please refer to implementation below for more info.

Structs

AbsoluteAndRelativeReserveProvider

MultiAsset reserve location provider. It’s based on RelativeReserveProvider and in addition will convert self absolute location to relative location.

AccountIdToMultiLocation

Convert AccountId to MultiLocation.

AllowTopLevelPaidExecutionFrom

Allows execution from origin if it is contained in T (i.e. T::Contains(origin)) taking payments into account.

AssetLocationIdConverter

Used to convert between cross-chain asset multilocation and local asset Id.

FixedRateOfForeignAsset

Used as weight trader for foreign assets.

ReserveAssetFilter

Used to determine whether the cross-chain asset is coming from a trusted reserve or not

XcmFungibleFeeHandler

Used to deposit XCM fees into a destination account.

Constants

MAX_ASSETS

XCM_SIZE_LIMIT